Move `mf_indent` to `string_helpers.py`.
