+++ metricflow-semantics/metricflow_semantics/query/group_by_item/group_by_item_resolver.py
@@ -10,7 +10,7 @@ from dbt_semantic_interfaces.references import SemanticModelReference, TimeDimen
 from dbt_semantic_interfaces.type_enums import TimeGranularity
 from typing_extensions import override
 
-from metricflow_semantics.mf_logging.formatting import mf_indent
+from metricflow_semantics.helpers.string_helpers import mf_indent
 from metricflow_semantics.mf_logging.lazy_formattable import LazyFormat
 from metricflow_semantics.mf_logging.pretty_print import mf_pformat
 from metricflow_semantics.model.semantic_manifest_lookup import SemanticManifestLookup
