+++ metricflow-semantics/metricflow_semantics/query/issues/filter_spec_resolver/invalid_where.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from dbt_semantic_interfaces.protocols import WhereFilter
 from typing_extensions import override
 
-from metricflow_semantics.mf_logging.formatting import mf_indent
+from metricflow_semantics.helpers.string_helpers import mf_indent
 from metricflow_semantics.query.group_by_item.resolution_path import MetricFlowQueryResolutionPath
 from metricflow_semantics.query.issues.issues_base import (
     MetricFlowQueryIssueType,
