+++ metricflow-semantics/metricflow_semantics/query/issues/group_by_item_resolver/no_common_items.py
@@ -5,7 +5,7 @@ from typing import Dict, Sequence, Tuple
 
 from typing_extensions import override
 
-from metricflow_semantics.mf_logging.formatting import mf_indent
+from metricflow_semantics.helpers.string_helpers import mf_indent
 from metricflow_semantics.mf_logging.pretty_print import mf_pformat
 from metricflow_semantics.naming.object_builder_scheme import ObjectBuilderNamingScheme
 from metricflow_semantics.query.group_by_item.candidate_push_down.group_by_item_candidate import GroupByItemCandidateSet
