+++ metricflow-semantics/metricflow_semantics/query/issues/group_by_item_resolver/no_matching_items_for_measure.py
@@ -6,7 +6,7 @@ from typing import Sequence, Tuple
 
 from typing_extensions import override
 
-from metricflow_semantics.mf_logging.formatting import mf_indent
+from metricflow_semantics.helpers.string_helpers import mf_indent
 from metricflow_semantics.mf_logging.pretty_print import mf_pformat
 from metricflow_semantics.query.group_by_item.resolution_path import MetricFlowQueryResolutionPath
 from metricflow_semantics.query.issues.issues_base import (
