+++ metricflow-semantics/metricflow_semantics/query/resolver_inputs/query_resolver_inputs.py
@@ -12,7 +12,7 @@ from typing import Optional, Tuple, Union
 from dbt_semantic_interfaces.protocols import WhereFilterIntersection
 from typing_extensions import override
 
-from metricflow_semantics.mf_logging.formatting import mf_indent
+from metricflow_semantics.helpers.string_helpers import mf_indent
 from metricflow_semantics.mf_logging.pretty_print import mf_pformat
 from metricflow_semantics.naming.metric_scheme import MetricNamingScheme
 from metricflow_semantics.naming.naming_scheme import QueryItemNamingScheme
